JGP-2: Must consider more than only head revs to identify branches for a given commit.
When notified by a Stash commit, Jenkins builds that exact commit regarless if it is still the head rev of a branch. When multiple notifications are send within a time period shorter than the job execution time, the built commit many not be the head rev of a branch anymore.