Fix FindBugs errors and warnings
With upgrade to plugin parent version 2.30, the FindBugs check is now
added to the Maven build lifecycle and we need to fix the FindBugs problems in order to be able to successfully finish the build.

Amongst others, FindBugs reported a problem for the class _PluginImpl_.
As the workaround for the classloading problems (like _HUDSON-4172_)
doesn't seem to be needed anymore, the class has simply been removed.